CAMBIOS 9-12-2025

+++ pages/proveedores/ordenCompra/updateOrdenCompra.js
@@ -187,6 +187,11 @@ const updateOrdenCompra = ({exito , ordenID}) => {
             bodyData.presupuesto = ordenCompra.presupuesto;
         }
 
+        if(ordenCompra.fechaEntrega===""){
+            alert("❌ Faltan completar algunos campos obligatorios.")
+            return
+        }
+
         const resOrdenCompra = await fetch(`${process.env.NEXT_PUBLIC_BACKEND_URL}/proveedor/ordenCompra/${ordenID}`, {
             method: 'PUT',
             headers: { 'Content-Type': 'application/json' },
@@ -194,9 +199,13 @@ const updateOrdenCompra = ({exito , ordenID}) => {
         })
 
         const ordenCompraCreado = await resOrdenCompra.json();
-        const ordenID = ordenCompraCreado.data._id;
+        if(!ordenCompraCreado.ok) {
+            alert(ordenCompraCreado.message)
+            return
+        }
+        const identificador = ordenCompraCreado.data._id;
 
-        await fetch(`${process.env.NEXT_PUBLIC_BACKEND_URL}/proveedor/comprobanteCompraDetalle/${ordenID}`,
+        await fetch(`${process.env.NEXT_PUBLIC_BACKEND_URL}/proveedor/ordenCompraDetalle/${identificador}`,
             {
                 method:'DELETE',
                 headers: {
@@ -205,10 +214,13 @@ const updateOrdenCompra = ({exito , ordenID}) => {
             }
         ).then((a)=>{return a.json()})
             .then((res)=>{
-                console.log(res.message);
+                if(!res.ok){
+                    alert(res.message)
+                    return
+                }
             })
             .catch((err)=>{
-                console.log("Error al enviar detalle de orden de compra para su eliminación. \n Error: ",err);
+                console.log("❌ Error al enviar detalle de orden de compra para su eliminación. \n Error: ",err);
             })
 
         // GUARDAMOS DETALLES
@@ -221,43 +233,34 @@ const updateOrdenCompra = ({exito , ordenID}) => {
                     precio: detalle.precio,
                     cantidad: detalle.cantidad,
                     importe: detalle.importe,
-                    ordenCompra: ordenID
+                    ordenCompra: identificador
             })
             
             
             });
-            if (!resDetalle.ok) throw new Error("Error al guardar un detalle");
-            
-            setDetalles([initialDetalle]);
-            setOrdenCompra(initialStateOrdenCompra);
-            exito();
+            if (!resDetalle.ok) {
+                const errorData = await resDetalle.json();
+                alert(errorData.message); 
+                return;
+            }
         }
+            
+        setDetalles([initialDetalle]);
+        setOrdenCompra(initialStateOrdenCompra);
+        alert(ordenCompraCreado.message)
+        exito();
     }
 
     const handleDetalleChange = (index, field, value) => {
         const nuevosDetalles = [...detalles];
-        nuevosDetalles[index][field] = field === "cantidad" ? parseFloat(value) : value;
-        
-        const prod = productos.find(p => p._id === nuevosDetalles[index].producto);
-       
-        if (prod) {
-            if(prod.precioCosto){
-                const ganancia = prod.ganancia;
-                const precio = prod.precioCosto + ((prod.precioCosto * ganancia) / 100);
-                
-                nuevosDetalles[index].precio = precio;
-                nuevosDetalles[index].importe = precio * nuevosDetalles[index].cantidad;
-            }
-            if(!prod.precioCosto && prod.precioVenta){
-                const precio = prod.precioVenta;
+        const detalle = nuevosDetalles[index];
 
-                nuevosDetalles[index].precio = precio;
-                nuevosDetalles[index].importe = precio * nuevosDetalles[index].cantidad;
-            }
+        // Actualizamos el campo directamente
+        detalle[field] = field === "cantidad" ? parseFloat(value) : value;
 
-        } else {
-            nuevosDetalles[index].precio = 0;
-            nuevosDetalles[index].importe = 0;
+        // Si cambió la cantidad → SOLO recalcular importe
+        if (field === "cantidad") {
+            detalle.importe = detalle.precio * detalle.cantidad;
         }
 
         setDetalles(nuevosDetalles);
@@ -615,10 +618,6 @@ const updateOrdenCompra = ({exito , ordenID}) => {
                         <div className="form-col-productos">
                             <label>
                                     Productos:
-                                    {/* <button type="button" className="btn-plus" onClick={() => setMostrarModalCreate3(true)}>+</button> */}
-                                    <button type="button" className="btn-add-producto" onClick={agregarDetalle}>
-                                        + Agregar Producto
-                                    </button>
                             </label>
                             <div className="form-group-presupuesto">
                                 
@@ -635,6 +634,7 @@ const updateOrdenCompra = ({exito , ordenID}) => {
                                             }
                                             placeholder="Tipo de Producto..."
                                             isClearable
+                                            isDisabled={true}
                                             styles={{
                                                 container: (base) => ({
                                                 ...base,
@@ -684,6 +684,7 @@ const updateOrdenCompra = ({exito , ordenID}) => {
                                             }
                                             placeholder="Producto..."
                                             isClearable
+                                            isDisabled={true}
                                             styles={{
                                                 container: (base) => ({
                                                 ...base,
@@ -782,7 +783,7 @@ const updateOrdenCompra = ({exito , ordenID}) => {
                                     className="submit-btn"
                                     onClick={(e) => {
                                         if (!puedeGuardar) {
-                                        alert("No se puede guardar una nota de pedido sin al menos un producto con cantidad.");
+                                        alert("❌ No se puede guardar una orden de compra sin al menos un producto con cantidad.");
                                         e.preventDefault();
                                         return;
                                         }
